Moved production ods url and loading url from config to utils from credential_service.cpp.
